requirements.txt
asgiref==3.6.0
Django==4.2
django-filter==23.2
djangorestframework==3.14.0
importlib-metadata==6.6.0
Markdown==3.4.3
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
zipp==3.15.0
